# src/main/resources/static/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lootwise API
  version: 1.0.0
  description: API spec for the Lootwise frontend.

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/v1/auth/signup:
    post:
      summary: Sign up
      operationId: signUp
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignUpResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"

  /api/v1/auth/login:
    post:
      summary: Login
      operationId: login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"

  /api/v1/users/me:
    get:
      summary: Get current user
      operationId: getMe
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserMeResponse"
        "401":
          description: Missing invalid or expired access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: User is withdrawn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"

  /api/v1/deals:
    get:
      summary: Get top discounted deals
      tags:
        - Deals
      operationId: getTopDeals
      parameters:
        - name: keyword
          in: query
          required: false
          schema:
            type: string
        - name: minDiscountPercent
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: maxFinalPrice
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: 0
        - name: minReviewScore
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: reviewScoreDesc
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/DealSort"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DealSummaryResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"

  /api/v1/deals/{appId}:
    get:
      summary: Get discounted deal detail by appId
      tags:
        - Deals
      operationId: getDealDetail
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealDetailResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Deal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SignUpRequest:
      type: object
      required:
        - email
        - password
        - nickname
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: test-user@example.com
        password:
          type: string
          minLength: 8
          maxLength: 72
          format: password
          example: Password123!
        nickname:
          type: string
          minLength: 2
          maxLength: 30
          pattern: "^[A-Za-z0-9가-힣_]+$"
          example: loot_user

    SignUpResponse:
      type: object
      required:
        - userId
        - email
        - nickname
      properties:
        userId:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: test-user@example.com
        nickname:
          type: string
          example: loot_user

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: test-user@example.com
        password:
          type: string
          maxLength: 72
          format: password
          example: Password123!

    LoginResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresInSeconds
      properties:
        accessToken:
          type: string
          example: jwt-access-token
        tokenType:
          type: string
          enum:
            - Bearer
          example: Bearer
        expiresInSeconds:
          type: integer
          format: int64
          example: 900

    UserMeResponse:
      type: object
      required:
        - userId
        - email
        - nickname
        - role
        - status
      properties:
        userId:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: test-user@example.com
        nickname:
          type: string
          example: loot_user
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"

    UserRole:
      type: string
      enum:
        - USER

    UserStatus:
      type: string
      enum:
        - ACTIVE
        - WITHDRAWN

    DealSummaryResponse:
      type: object
      required:
        - appId
        - name
        - finalPrice
        - discountPercent
        - steamUrl
      properties:
        appId:
          type: integer
          format: int64
        name:
          type: string
        originalPrice:
          type: number
          format: double
          nullable: true
        finalPrice:
          type: number
          format: double
        discountPercent:
          type: integer
        reviewScoreDesc:
          type: string
          nullable: true
        reviewScoreDescKo:
          type: string
          nullable: true
        steamUrl:
          type: string
        capsuleImageUrl:
          type: string
          nullable: true

    DealDetailResponse:
      allOf:
        - $ref: "#/components/schemas/DealSummaryResponse"

    ApiErrorResponse:
      type: object
      required:
        - message
        - timestamp
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time

    DealSort:
      type: string
      enum:
        - DISCOUNT_DESC
        - PRICE_ASC
        - REVIEW_COUNT_DESC
        - REVIEW_SCORE_DESC
        - LATEST_DESC